expose range

+++ include/txtvm/expr.h
@@ -92,7 +92,7 @@ namespace txtvm {
 
 
     Expr IntConstant(int64_t value);
-    Expr FloatConstant(int64_t value);
+    Expr FloatConstant(double value);
 
     /*! \brief base of expression node */
     class ExprNode : public Node {
